adding search movie query to core

+++ src/House.Flix.Models/Common/PagedResult.cs
@@ -1,3 +1,3 @@
 namespace House.Flix.Models.Common;
 
-public record PagedResult<T>(T[] Items, long Page, long PageSize, long TotalCount);
+public record PagedResult<T>(T[] Items, int Page, int PageSize, int TotalCount);
